Drop custom MP interface
Custom interface turned out to be too cumbersome to use. Now or_tools
MPSolver is used directly in MP formulations. This allows full
leverage of or_tools interface, which makes the code easier to
understand and less prone to bugs.

TEST ARE FAILING after this change and ESCF is not working properly.